refactor: 폴더구조 변경 및 네이밍 수정, userDecorator 타입 지정
