+++ src/auth/utils/RTokenStrategy.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { JwtPayload } from 'src/utils/types';
+import { JwtPayload } from 'src/types/jwt.interface';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { Request } from 'express';
 import { ConfigService } from '@nestjs/config';
