+++ src/types/jwt.interface.ts
@@ -1,0 +1,5 @@
+export type JwtPayload = {
+  id: number;
+  email: string;
+  refreshToken?: string;
+};
