+++ src/utils/user.decorator.ts
@@ -1,5 +1,6 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-export const UserDecorator = createParamDecorator(
+import UserAuthInterface from '../types/user.interface';
+export const UserDecorator = createParamDecorator<UserAuthInterface>(
   (data: unknown, ctx: ExecutionContext) => {
     const request = ctx.switchToHttp().getRequest();
     return request.user;
